requirements.txt
annotated-types==0.7.0
azure-core==1.31.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
filelock==3.16.1
fsspec==2024.9.0
ginza==5.2.0
ginza-transformers==0.4.2
huggingface-hub==0.25.2
idna==3.10
ja-ginza-electra==5.2.0
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
logzero==1.7.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.1
numpy==2.0.2
packaging==24.1
phonenumbers==8.13.47
plac==1.4.3
preshed==3.0.9
presidio_analyzer==2.2.355
presidio_anonymizer==2.2.355
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rich==13.9.2
safetensors==0.4.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.1.9
srsly==2.4.8
SudachiDict-core==20240716
SudachiPy==0.6.8
SudachiTra==0.1.9
sympy==1.13.3
thinc==8.3.2
tldextract==5.1.2
tokenizers==0.13.3
torch==2.4.1
tqdm==4.66.5
transformers==4.25.1
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
